finally fix all test errors (?)
