+++ requirements_test.txt
@@ -1,1 +1,1 @@
-rpy2>=3.5.1
+rpy2>=3.5.3
